add triple buffering for sensor data storage

+++ common/all/interfaces/include/interfaces/IClockSource.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 namespace interfaces
 {
 
@@ -8,17 +10,17 @@ concept IClockSource = requires {
    {
       T::now_us()
    }
-   -> std::same_as<uint64_t>;
+   -> std::same_as<uint32_t>;
 
    {
       T::now_ms()
    }
-   -> std::same_as<uint64_t>;
+   -> std::same_as<uint32_t>;
 
    {
       T::now_s()
    }
-   -> std::same_as<uint64_t>;
+   -> std::same_as<uint32_t>;
 };
 
 }   // namespace interfaces
